+++ alasql.js
@@ -4246,8 +4246,8 @@ yy.Op.prototype.toJavaScript = function(context,tableid) {
 	else if(this.op == 'OR') op = '||';
 
 	if(this.op == 'LIKE') {
-		var s = "('^'"+this.left.toJavaScript(context,tableid)+")"+
-		".match(new RegExp(("+this.right.toJavaScript(context,tableid)+").replace(/\\\%/g,'.*'),'g')+'$')"
+		var s = "("+this.left.toJavaScript(context,tableid)+")"+
+		".match(new RegExp('^'+("+this.right.toJavaScript(context,tableid)+").replace(/\\\%/g,'.*')+'$','g'))"
 //		console.log(s);
 		return s;
 	};