+++ test/test110.js
@@ -24,7 +24,7 @@ describe('LEFT JOIN', function(){
 		alasql.exec('INSERT INTO test1 VALUES (1,5)');
 		alasql.exec('INSERT INTO test1 VALUES (2,6)');
 
-		var res = alasql.exec('SELECT SUM(b) AS sb,a,test1.c FROM test LEFT JOIN test1 ON test.a = test1.a GROUP BY c,a');
+		var res = alasql.exec('SELECT SUM(b) AS sb,test1.a,test1.c FROM test LEFT JOIN test1 ON test.a = test1.a GROUP BY c,a');
 
 		assert.equal(8, res[0].sb);
 		assert.equal(2, res[1].sb);